requirements.txt
djangorestframework
django-filter
psycopg2
djangorestframework-simplejwt
dotenv